gbx: store sample floats as f32 -- the width the file actually has
The decoder held every sample float as f64. The record does not: position and
velocity are f32, and the orientation is three PACKED INTEGERS (a u16 rotation
and two i16 angles) carrying about five decimal digits. The f64 was a port
artifact -- the original implementation was Python, where every float is f64.

`Sample`'s floats are now f32. The trig still runs in f64 (intermediate
precision is worth having) and narrows once, at the boundary, via a single `n`
closure in `decode_vehicle_sample`. ~20 files across ghost, tmtraj, tmauto,
clip, resim and mapgeom widen with `as f64` where they accumulate.

WHAT IT FIXES, measured on real data before the refactor was started: decoding
p00001_19538 in debug and in release produced JSON differing in 3 bytes of
454 KB, and narrowing every float in both dumps to f32 made them identical.
The golden now holds ACROSS PROFILES -- blessed under --release, passes under
debug, which it never did before:

    release  ok. 2 passed
    debug    ok. 2 passed

`Val::F` is f32 too, and `json::py_repr_f32` renders it WITHOUT widening: the
shortest string that round-trips as f64 is `0.10000000149011612` where the same
value as f32 is `0.1`. Widening would have kept the ugly output and thrown away
the whole point.

WHAT IT COSTS, and this is a real cost that needs a decision -- see the next
commit message. `reproduces_the_51_python_trajectories` now fails on all 51
runs. That test asserts the Rust port reproduces the original Python's rendered
output EXACTLY ("the 'rendered' column is what the port must reproduce, and it
is exactly zero"). The difference is one digit in the sixth significant figure
of derived angles:

    python  ...,0.00826859,0.000929603,...
    rust    ...,0.00826859,0.000929604,...

f32 carries ~7 significant digits, so rounding to the CSV's 6 can flip the
last one. Python computed in f64 and our f64 reproduced it exactly; f32 cannot.

Full suite: 462 passed, 2 failed -- the 51-trajectory test above, and the
pre-existing clip/ffmpeg WSL-path test.

+++ tools/clip/src/overlay.rs
@@ -612,7 +612,7 @@ pub fn alignment(ghost: &str, span_ms: i64) -> Result<(i64, f64, f64), String> {
             let Some(i) = ins.get((t / 10) as usize) else { continue };
             // The record stores steer as `floor((steer_i8 + 127) * 255 / 254)`;
             // comparing in i8 space avoids re-deriving that here.
-            err += (s.steer * 127.0 - i.steer as f64).abs();
+            err += (s.steer as f64 * 127.0 - i.steer as f64).abs();
             n += 1;
         }
         if n > 20 {
